Fix bug #322 - Installing with composer --no-dev
After installing and generating a blog, there is a phase for
deploying it. It is used normally with composer the tag --no-dev
to avoid installing development tools that are not required on a
production environment.

However, the autosetup tools of canvas require the use of faker
library so update canvas environment.

This commit just moves faker from "require-dev" to "require" so
that the toolchain is not broken.